refactor(audio_analysis): the grid constant belonged to the grid, not to the axes
`DEFAULT_TIME_GRID` was declared in `axes.py` and `axes.py` never read it: it
appeared there once, as its own assignment target, and every other mention in
that file is prose. `grid.py` imported it back the other way to use as
`BucketGrid`'s default. That inverted the arrow — an axis *reads* the grid it is
estimated on, so the grid does not belong to the axes — and the single line it
took to import it back was, at import time, the **only** path from the
extraction layer to the refiner's axis vocabulary.

Measured before the move, by AST over the package and confirmed against
`sys.modules` in a fresh interpreter: the closure of `stages.py` +
`stage_context.py` is 8 modules / 2,939 lines at import time, and the one edge to
`axes` is `stages` -> `sound_sources` -> `grid` -> `axes`, carrying this one
constant. `calibration` and `types` also reach `axes`, but only through
function-local imports, so neither was ever on the import-time path.

The constant and its docstring now live above `BucketGrid`, with that reasoning
recorded beside them. `grid.py` ends up with **zero** intra-package imports,
which is what makes the boundary checkable rather than incidental; its module
docstring says so, so the next edit that adds one has to argue with it.

Every reader re-pointed: `run_config`, `adaptive/loop`, and the four test sites.
The prose references in `asr`, `compute`, `doc.md`, `cached_inference` and the
config's own derivation block are renamed to match — a `:data:` reference that
resolves to nothing is the same class of stale as a wrong one. The spec files
under `specs/` keep `axes.DEFAULT_TIME_GRID`: they are dated records of what was
read at the time.
